+++ packages/webapp/src/stories/Form/radio/radio.scss
@@ -3,69 +3,65 @@
 .container {
   display: inline-block;
   position: relative;
-  padding-left: 25px;
+  padding-left: 24px;
   cursor: pointer;
+  font-size: 22px;
   -webkit-user-select: none;
   -moz-user-select: none;
   -ms-user-select: none;
   user-select: none;
 }
 
-.label {
-  @include form-font();
-  transform: translateY(-2px);
-  margin-bottom: 0;
-}
-
-/* Hide the browser's default checkbox */
+/* Hide the browser's default radio button */
 .container input {
   position: absolute;
   opacity: 0;
   cursor: pointer;
-  height: 0;
-  width: 0;
 }
 
-/* Create a custom checkbox */
+/* Create a custom radio button */
 .checkmark {
   position: absolute;
   top: 0;
   left: 0;
-  height: 18px;
-  width: 18px;
+  height: 16px;
+  width: 16px;
   background-color: white;
+  border-radius: 50%;
   border: solid 2px var(--iconDefault);
-  border-radius: 2px;
-}
-// Change the color of the checkbox
-/* When the checkbox is checked, add a green background */
-.container input:checked ~ .checkmark {
-  background-color: var(--checkbox);
-  border: none;
 }
 
-/* Create the checkmark/indicator (hidden when not checked) */
+/* When the radio button is checked, add a blue background */
+.container input:checked ~ .checkmark {
+  border: solid 2px var(--checkbox);
+}
+
+/* Create the indicator (the dot/circle - hidden when not checked) */
 .checkmark:after {
   content: "";
   position: absolute;
   display: none;
 }
 
-/* Show the checkmark when checked */
+/* Show the indicator (dot/circle) when checked */
 .container input:checked ~ .checkmark:after {
   display: block;
 }
 
-/* Style the checkmark/indicator */
+/* Style the indicator (dot/circle) */
 .container .checkmark:after {
-  left: 5px;
-  width: 7px;
-  height: 13px;
-  border: solid white;
-  border-width: 0 2px 2px 0;
-  -webkit-transform: rotate(45deg);
-  -ms-transform: rotate(45deg);
-  transform: rotate(45deg);
+  top: 2px;
+  left: 2px;
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  background-color: var(--checkbox);
+}
+
+.label {
+  @include form-font();
+  transform: translateY(-4px);
+  margin-bottom: 0;
 }
 
 .disabled{
