+++ packages/webapp/src/stories/assets/colors.scss
@@ -47,8 +47,15 @@
   --inputActive: var(--teal500);
   --fontColor: var(--grey900);
   --border: var(--grey200);
+  --checkbox: var(--teal700);
 }
 
 @mixin gradient(){
   background: linear-gradient(96.68deg, #78C99E -4.29%, #C7EFD3 24.32%, #E3F8EC 35.52%, #E3F8EC 64.28%, #C7EFD3 80.81%, #78C99E 125.09%);
+}
+
+@mixin form-font(){
+  font-size: 16px;
+  line-height: 24px;
+  color: var(--fontColor);
 }
